re-org simulation tests
