+++ .circleci/config.yml
@@ -123,7 +123,7 @@ jobs:
       - run:
           name: Go Test with Coverage
           no_output_timeout: 2400
-          command: go test $(glide nv | grep -v Utilities | grep -v LongTests  | grep -v simTest) -vet=off -v -cover -coverprofile=coverage.out
+          command: go test $(glide nv | grep -v Utilities | grep -v longTest | grep -v peerTest) -vet=off -v -cover -coverprofile=coverage.out
 
       - run:
           name: Coveralls!
