+++ longTest/README.md
@@ -1,0 +1,4 @@
+# factomd/longTest
+
+This folder contains simulation tests that take a very long time to run.
+These tests are *not run* on circle.ci
