+++ peerTest/README.md
@@ -1,0 +1,8 @@
+# factomd/peerTest
+
+This folder contains tests that must be run in parallel.
+
+These tests are useful for testing features between builds
+by running 1 of each pair from previous/current builds.
+
+Tests in this folder will *not run* on circle.ci
