+++ simTest/README.md
@@ -1,0 +1,4 @@
+# factomd/simTest
+
+This folder contains simulation tests that can run alone in isolation.
+Tests in this folder will also run on circle.ci
