+++ support/dev/simulator/brainSwap/test1.sh
@@ -1,6 +1,5 @@
 #!/usr/bin/env bash
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
-SIM_TEST=${DIR}/../../../../simTest/BrainSwapFollower_test.go
-cd ${DIR}/v1
-go test -v $SIM_TEST #> out1.txt
+SIM_TEST=${DIR}/../../../../peerTest/BrainSwapFollower_test.go
+cd ${DIR}/v1 && go test -v $SIM_TEST #> out1.txt
